minecraftrolesync/requirements.txt
discord.py==1.7.3
pyyaml==5.4.1
mcstatus==6.0.1
